train with eval

+++ tools/calculate_fid.py
@@ -508,11 +508,11 @@ def get_activations(
             if pred.size(2) != 1 or pred.size(3) != 1:
                 pred = adaptive_avg_pool2d(pred, output_size=(1, 1))
 
-            pred = pred.squeeze(3).squeeze(2).cpu().numpy()
+            pred = pred.squeeze(3).squeeze(2).float().cpu().numpy()
         else:
             with torch.no_grad():
                 pred = model(batch)
-            pred = pred.cpu().numpy()
+            pred = pred.float().cpu().numpy()
 
         pred_arr[start_idx : start_idx + pred.shape[0]] = pred
